Code review fixes
