+++ example/e2e/sanityTest.spec.js
@@ -17,17 +17,4 @@ describe('NetInfo', () => {
   it('should load example app with no errors and show all the examples by default', async () => {
     await expect(element(by.id('examplesTitle'))).toExist();
   });
-
-  it('should show the test case when opening a URL', async () => {
-    await device.openURL({url: 'netinfoexample://emitOnListen'});
-
-    await expect(element(by.id('testCasesTitle'))).toExist();
-  });
-
-  it('should clear the test case when asked to', async () => {
-    await device.openURL({url: 'netinfoexample://emitOnListen'});
-    await device.openURL({url: 'netinfoexample://clear'});
-
-    await expect(element(by.id('examplesTitle'))).toExist();
-  });
 });
