+++ example/testCases/DeprecatedMultipleGetConnectionInfo.tsx
@@ -9,7 +9,7 @@
 
 import * as React from 'react';
 import {Button, Text, View} from 'react-native';
-import NetInfo from '../../js';
+import NetInfo from '../../src';
 
 interface State {
   results: boolean[];
