+++ jest.config.js
@@ -8,7 +8,7 @@ module.exports = {
     ...tsjPreset.transform,
     '\\.js$': '<rootDir>/node_modules/react-native/jest/preprocessor.js',
   },
-  testPathIgnorePatterns: ['<rootDir>/lib/', '<rootDir>/node_modules/'],
+  testPathIgnorePatterns: ['<rootDir>/lib/', '<rootDir>/node_modules/', '<rootDir>/example/'],
   globals: {
     'ts-jest': {
       babelConfig: true,
