+++ src/__tests__/deprecated/getConnectionInfo.ts
