+++ src/__tests__/deprecated/isConnected.ts
