+++ src/__tests__/deprecated/isConnectionExpensiveAndroid.ts
