+++ src/__tests__/deprecated/isConnectionExpensiveIOS.ts
