+++ src/internal/__tests__/deprecatedUtils.ts
