+++ src/internal/deprecatedUtils.ts
