+++ src/internal/nativeInterface.ts
