+++ src/internal/privateTypes.ts
