+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["js/**/*.ts", "example/**/*.ts", "example/**/*.tsx"],
+  "include": ["src/**/*.ts", "example/**/*.ts", "example/**/*.tsx"],
   "compilerOptions": {
     "target": "es6",
     "module": "es6",
